+++ mipengine/udfgen/__init__.py
@@ -6,11 +6,11 @@ from mipengine.udfgen.iotypes import literal
 from mipengine.udfgen.iotypes import merge_tensor
 from mipengine.udfgen.iotypes import merge_transfer
 from mipengine.udfgen.iotypes import relation
-from mipengine.udfgen.iotypes import secure_transfer
 from mipengine.udfgen.iotypes import state
 from mipengine.udfgen.iotypes import tensor
 from mipengine.udfgen.iotypes import transfer
 from mipengine.udfgen.iotypes import udf_logger
+from mipengine.udfgen.smpc import secure_transfer
 from mipengine.udfgen.udfgenerator import generate_udf_queries
 
 __all__ = [
