+++ mipengine/udfgen/iotypes.py
@@ -128,7 +128,7 @@ class TensorType(TableType, ParametrizedType, InputType, OutputType):
         columns_tmpl = "{{name: _columns[name_w_prefix] for name, name_w_prefix in zip({colnames}, {colnames_w_prefix})}}"
         return f"{{varname}} = udfio.from_tensor_table({columns_tmpl})"
 
-    def get_main_return_stmt_template(self, _) -> str:
+    def get_main_return_stmt_template(self) -> str:
         return "return udfio.as_tensor_table(numpy.array({return_name}))"
 
 
@@ -201,7 +201,7 @@ class RelationType(TableType, ParametrizedType, InputType, OutputType):
         columns_tmpl = "{{name: _columns[name_w_prefix] for name, name_w_prefix in zip({colnames}, {colnames_w_prefix})}}"
         return f"{{varname}} = udfio.from_relational_table({columns_tmpl}, '{ROWID}')"
 
-    def get_main_return_stmt_template(self, _) -> str:
+    def get_main_return_stmt_template(self) -> str:
         return f"return udfio.as_relational_table({{return_name}}, '{ROWID}')"
 
 
@@ -240,10 +240,10 @@ class TransferType(DictType, InputType, LoopbackOutputType):
             ]
         )
 
-    def get_main_return_stmt_template(self, _) -> str:
+    def get_main_return_stmt_template(self) -> str:
         return "return json.dumps({return_name})"
 
-    def get_secondary_return_stmt_template(self, tablename_placeholder, _) -> str:
+    def get_secondary_return_stmt_template(self, tablename_placeholder) -> str:
         return (
             '_conn.execute(f"INSERT INTO $'
             + tablename_placeholder
@@ -273,184 +273,12 @@ def merge_transfer():
     return MergeTransferType()
 
 
-class SecureTransferType(DictType, InputType, LoopbackOutputType):
-    _data_column_name = "secure_transfer"
-    _data_column_type = dt.JSON
-    _sum_op: bool
-    _min_op: bool
-    _max_op: bool
-
-    def __init__(self, sum_op=False, min_op=False, max_op=False):
-        self._sum_op = sum_op
-        self._min_op = min_op
-        self._max_op = max_op
-
-    @property
-    def sum_op(self):
-        return self._sum_op
-
-    @property
-    def min_op(self):
-        return self._min_op
-
-    @property
-    def max_op(self):
-        return self._max_op
-
-    def get_build_template(self):
-        colname = self.data_column_name
-        return LN.join(
-            [
-                f'__transfer_strs = _conn.execute("SELECT {colname} from {{table_name}};")["{colname}"]',
-                "__transfers = [json.loads(str) for str in __transfer_strs]",
-                "{varname} = udfio.secure_transfers_to_merged_dict(__transfers)",
-            ]
-        )
-
-    def get_main_return_stmt_template(self, smpc_used) -> str:
-        return self._get_secure_transfer_main_return_stmt_template(smpc_used)
-
-    def _get_secure_transfer_main_return_stmt_template(self, smpc_used):
-        if smpc_used:
-            return_stmts = [
-                "template, sum_op, min_op, max_op = udfio.split_secure_transfer_dict({return_name})"
-            ]
-            (
-                _,
-                sum_op_tmpl,
-                min_op_tmpl,
-                max_op_tmpl,
-            ) = _get_smpc_table_template_names(MAIN_TABLE_PLACEHOLDER)
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.sum_op, sum_op_tmpl, "sum_op"
-                )
-            )
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.min_op, min_op_tmpl, "min_op"
-                )
-            )
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.max_op, max_op_tmpl, "max_op"
-                )
-            )
-            return_stmts.append("return json.dumps(template)")
-            return LN.join(return_stmts)
-        else:
-            # Treated as a TransferType
-            return "return json.dumps({return_name})"
-
-    def get_secondary_return_stmt_template(
-        self, tablename_placeholder, smpc_used
-    ) -> str:
-        return self._get_secure_transfer_sec_return_stmt_template(
-            tablename_placeholder, smpc_used
-        )
-
-    def _get_secure_transfer_sec_return_stmt_template(
-        self, tablename_placeholder: str, smpc_used
-    ):
-        if smpc_used:
-            return_stmts = [
-                "template, sum_op, min_op, max_op = udfio.split_secure_transfer_dict({return_name})"
-            ]
-            (
-                template_tmpl,
-                sum_op_tmpl,
-                min_op_tmpl,
-                max_op_tmpl,
-            ) = _get_smpc_table_template_names(tablename_placeholder)
-            return_stmts.append(
-                '_conn.execute(f"INSERT INTO $'
-                + template_tmpl
-                + " VALUES ('{{json.dumps(template)}}');\")"
-            )
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.sum_op, sum_op_tmpl, "sum_op"
-                )
-            )
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.min_op, min_op_tmpl, "min_op"
-                )
-            )
-            return_stmts.extend(
-                self._get_secure_transfer_op_return_stmt_template(
-                    self.max_op, max_op_tmpl, "max_op"
-                )
-            )
-            return LN.join(return_stmts)
-        else:
-            # Treated as a TransferType
-            return (
-                '_conn.execute(f"INSERT INTO $'
-                + tablename_placeholder
-                + " VALUES ('{{json.dumps({return_name})}}');\")"
-            )
-
-    @staticmethod
-    def _get_secure_transfer_op_return_stmt_template(
-        op_enabled, table_name_tmpl, op_name
-    ):
-        if not op_enabled:
-            return []
-        return [
-            '_conn.execute(f"INSERT INTO $'
-            + table_name_tmpl
-            + f" VALUES ('{{{{json.dumps({op_name})}}}}');\")"
-        ]
-
-    def get_smpc_build_template(self):
-        def get_smpc_op_template(enabled, operation_name):
-            stmts = []
-            if enabled:
-                stmts.append(
-                    f'__{operation_name}_values_str = _conn.execute("SELECT secure_transfer from {{{operation_name}_values_table_name}};")["secure_transfer"][0]'
-                )
-                stmts.append(
-                    f"__{operation_name}_values = json.loads(__{operation_name}_values_str)"
-                )
-            else:
-                stmts.append(f"__{operation_name}_values = None")
-            return stmts
-
-        stmts = []
-        stmts.append(
-            '__template_str = _conn.execute("SELECT secure_transfer from {template_table_name};")["secure_transfer"][0]'
-        )
-        stmts.append("__template = json.loads(__template_str)")
-        stmts.extend(get_smpc_op_template(self.sum_op, "sum_op"))
-        stmts.extend(get_smpc_op_template(self.min_op, "min_op"))
-        stmts.extend(get_smpc_op_template(self.max_op, "max_op"))
-        stmts.append(
-            "{varname} = udfio.construct_secure_transfer_dict(__template,__sum_op_values,__min_op_values,__max_op_values)"
-        )
-        return LN.join(stmts)
+class TransferTypeBase(ABC):
+    pass
 
 
-def _get_smpc_table_template_names(prefix: str):
-    """
-    This is used when a secure transfer is returned with smpc enabled.
-    The secure_transfer is one output_type but needs to be broken into
-    multiple tables, hence more than one main table names are needed.
-    """
-    return (
-        prefix,
-        prefix + "_sum_op",
-        prefix + "_min_op",
-        prefix + "_max_op",
-    )
-
-
-def secure_transfer(sum_op=False, min_op=False, max_op=False):
-    if not sum_op and not min_op and not max_op:
-        raise ValueError(
-            "In a secure_transfer at least one operation should be enabled."
-        )
-    return SecureTransferType(sum_op, min_op, max_op)
+TransferTypeBase.register(TransferType)
+TransferTypeBase.register(MergeTransferType)
 
 
 class StateType(DictType, InputType, LoopbackOutputType):
@@ -466,10 +294,10 @@ class StateType(DictType, InputType, LoopbackOutputType):
             ]
         )
 
-    def get_main_return_stmt_template(self, _) -> str:
+    def get_main_return_stmt_template(self) -> str:
         return "return pickle.dumps({return_name})"
 
-    def get_secondary_return_stmt_template(self, tablename_placeholder, _) -> str:
+    def get_secondary_return_stmt_template(self, tablename_placeholder) -> str:
         return (
             '_conn.execute(f"INSERT INTO $'
             + tablename_placeholder
@@ -594,43 +422,6 @@ class TransferArg(DictArg):
 
     def __init__(self, table_name: str):
         super().__init__(table_name)
-
-
-class SecureTransferArg(DictArg):
-    type = SecureTransferType()
-
-    def __init__(self, table_name: str):
-        super().__init__(table_name)
-
-
-class SMPCSecureTransferArg(UDFArgument):
-    type: SecureTransferType
-    template_table_name: str
-    sum_op_values_table_name: str
-    min_op_values_table_name: str
-    max_op_values_table_name: str
-
-    def __init__(
-        self,
-        template_table_name: str,
-        sum_op_values_table_name: str,
-        min_op_values_table_name: str,
-        max_op_values_table_name: str,
-    ):
-        sum_op = False
-        min_op = False
-        max_op = False
-        if sum_op_values_table_name:
-            sum_op = True
-        if min_op_values_table_name:
-            min_op = True
-        if max_op_values_table_name:
-            max_op = True
-        self.type = SecureTransferType(sum_op, min_op, max_op)
-        self.template_table_name = template_table_name
-        self.sum_op_values_table_name = sum_op_values_table_name
-        self.min_op_values_table_name = min_op_values_table_name
-        self.max_op_values_table_name = max_op_values_table_name
 
 
 class LiteralArg(UDFArgument):
