+++ tests/standalone_tests/udfgen/test_decorator.py
@@ -6,11 +6,11 @@ from mipengine.udfgen.decorator import UDFBadDefinition
 from mipengine.udfgen.decorator import udf
 from mipengine.udfgen.iotypes import merge_transfer
 from mipengine.udfgen.iotypes import relation
-from mipengine.udfgen.iotypes import secure_transfer
 from mipengine.udfgen.iotypes import state
 from mipengine.udfgen.iotypes import tensor
 from mipengine.udfgen.iotypes import transfer
 from mipengine.udfgen.iotypes import udf_logger
+from mipengine.udfgen.smpc import secure_transfer
 
 
 class TestUDFValidation:
